back_end/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dbapi==0.0.14
Flask==2.2.2
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
idna==3.4
inflect==6.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
Pillow==9.3.0
pydantic==1.10.2
PyMySQL==1.0.2
requests==2.28.1
sqlacodegen==2.3.0
SQLAlchemy==1.4.42
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
yapf==0.32.0
